refactor(models): split models/wow.py into guild/crafting/characters subpackage
884-line monolith split into three focused files under models/wow/.
The __init__.py re-exports all public symbols so no import sites need changes.
Also updates the TTL-expiry test monkeypatch target to the canonical module path.
