+++ NerdyPy/models/wow/characters.py
@@ -1,0 +1,66 @@
+# -*- coding: utf-8 -*-
+"""WoW character-related database models."""
+
+from datetime import UTC
+
+from sqlalchemy import (
+    Column,
+    DateTime,
+    ForeignKey,
+    Index,
+    Integer,
+    String,
+    Text,
+    Unicode,
+)
+from utils import database as db
+
+
+class WowCharacterMounts(db.BASE):
+    """Stored mount set per player (account-wide, keyed by highest-level char)."""
+
+    __tablename__ = "WowCharacterMounts"
+    __table_args__ = (
+        Index("WowCharacterMounts_ConfigId", "ConfigId"),
+        Index("WowCharacterMounts_Config_Char", "ConfigId", "CharacterName", "RealmSlug", unique=True),
+    )
+
+    Id = Column(Integer, primary_key=True)
+    ConfigId = Column(Integer, ForeignKey("WowGuildNewsConfig.Id"))
+    CharacterName = Column(Unicode(50))
+    RealmSlug = Column(String(100))
+    KnownMountIds = Column(Text, default="[]")
+    LastChecked = Column(DateTime, nullable=True)
+
+    @classmethod
+    def get_by_character(cls, config_id, char_name, realm_slug, session):
+        return (
+            session.query(cls)
+            .filter(cls.ConfigId == config_id)
+            .filter(cls.CharacterName == char_name)
+            .filter(cls.RealmSlug == realm_slug)
+            .first()
+        )
+
+    @classmethod
+    def get_all_by_config(cls, config_id, session):
+        return session.query(cls).filter(cls.ConfigId == config_id).all()
+
+    @classmethod
+    def delete_stale(cls, config_id, active_keys, stale_cutoff, session):
+        """Delete entries for characters no longer in the roster whose LastChecked is older than stale_cutoff.
+
+        active_keys: set of (CharacterName, RealmSlug) currently in the guild roster.
+        Returns the number of deleted entries.
+        """
+        all_entries = cls.get_all_by_config(config_id, session)
+        deleted = 0
+        for entry in all_entries:
+            if (entry.CharacterName, entry.RealmSlug) not in active_keys:
+                last_checked = entry.LastChecked
+                if last_checked and last_checked.tzinfo is None:
+                    last_checked = last_checked.replace(tzinfo=UTC)
+                if last_checked and last_checked < stale_cutoff:
+                    session.delete(entry)
+                    deleted += 1
+        return deleted
