+++ NerdyPy/models/wow/guild.py
@@ -1,0 +1,106 @@
+# -*- coding: utf-8 -*-
+"""WoW guild-related database models."""
+
+from datetime import UTC, datetime
+
+from sqlalchemy import (
+    BigInteger,
+    Boolean,
+    Column,
+    DateTime,
+    Index,
+    Integer,
+    String,
+    Text,
+    Unicode,
+    asc,
+)
+from utils import database as db
+
+
+class WoW(db.BASE):
+    """Database entity model for World of Warcraft"""
+
+    __tablename__ = "WoW"
+    __table_args__ = (Index("WoW_GuildId", "GuildId"),)
+
+    GuildId = Column(BigInteger, primary_key=True)
+
+
+class WowGuildNewsConfig(db.BASE):
+    """Tracks which Discord guilds monitor which WoW guilds for news."""
+
+    __tablename__ = "WowGuildNewsConfig"
+    __table_args__ = (
+        Index("WowGuildNewsConfig_GuildId", "GuildId"),
+        Index("WowGuildNewsConfig_Id_GuildId", "Id", "GuildId", unique=True),
+        Index(
+            "WowGuildNewsConfig_Guild_Realm_Region",
+            "GuildId",
+            "WowGuildName",
+            "WowRealmSlug",
+            "Region",
+            unique=True,
+        ),
+    )
+
+    Id = Column(Integer, primary_key=True)
+    GuildId = Column(BigInteger)
+    ChannelId = Column(BigInteger)
+    WowGuildName = Column(Unicode(100))
+    WowGuildNameDisplay = Column(Unicode(100), nullable=True)
+    WowRealmSlug = Column(String(100))
+    Region = Column(String(10))
+    Language = Column(String(5), default="en")
+    MinLevel = Column(Integer, default=10)
+    ActiveDays = Column(Integer, default=7)
+    RosterOffset = Column(Integer, default=0)
+    LastActivityTimestamp = Column(DateTime, nullable=True)
+    Enabled = Column(Boolean, default=True)
+    CreateDate = Column(DateTime, default=lambda: datetime.now(UTC))
+    AccountGroupData = Column(Text, default="{}")
+
+    @classmethod
+    def get_by_id(cls, config_id, guild_id, session):
+        return session.query(cls).filter(cls.Id == config_id).filter(cls.GuildId == guild_id).first()
+
+    @classmethod
+    def get_all_enabled(cls, session):
+        return session.query(cls).filter(cls.Enabled.is_(True)).order_by(asc("Id")).all()
+
+    @classmethod
+    def get_all_by_guild(cls, guild_id, session):
+        return session.query(cls).filter(cls.GuildId == guild_id).order_by(asc("Id")).all()
+
+    @classmethod
+    def get_existing(cls, guild_id, wow_guild_name, realm_slug, region, session):
+        """Check if a config already exists for this guild+realm+region combination."""
+        return (
+            session.query(cls)
+            .filter(
+                cls.GuildId == guild_id,
+                cls.WowGuildName == wow_guild_name,
+                cls.WowRealmSlug == realm_slug,
+                cls.Region == region,
+            )
+            .first()
+        )
+
+    @classmethod
+    def delete(cls, config_id, guild_id, session):
+        from models.wow.characters import WowCharacterMounts
+
+        config = cls.get_by_id(config_id, guild_id, session)
+        if config:
+            session.query(WowCharacterMounts).filter(WowCharacterMounts.ConfigId == config.Id).delete()
+            session.delete(config)
+
+    def __str__(self):
+        status = "active" if self.Enabled else "paused"
+        return (
+            f"==== {self.Id} ====\n"
+            f"Guild: {self.WowGuildName} ({self.WowRealmSlug}-{self.Region.upper()})\n"
+            f"Language: {self.Language}\n"
+            f"Status: {status}\n"
+            f"Active Days Filter: {self.ActiveDays}\n"
+        )
